write script that fixes wrong job namespace
